debian package: add missing git build dependency
pbr uses git to derive the python package version from the git tag

without this the build fails with:

Versioning for this project requires either an sdist tarball, or access to an upstream git repository. It's also possible that there is a mismatch between the package name in setup.cfg and the argument given to pbr.version.VersionInfo. Project name virtualbmc was given, but was not able to be found.